moved implementation of SendInputSensorControlIndication from osp-api.c to sensoracq_t.c created header file sensoracq_t.h
