+++ embedded/common/app/osp-api.c
@@ -24,6 +24,7 @@
 #include "osp_embeddedforegroundalgcalls.h"
 #include "osp_embeddedbackgroundalgcalls.h"
 #include "osp-version.h"
+#include "sensoracq_t.h"
 
 
 /*-------------------------------------------------------------------------------------------------*\
@@ -646,26 +647,6 @@ void InvalidateQueuedDataByHandle(InputSensorHandle_t Handle)
 	}
 	ExitCritical();
 }
-
-
-/****************************************************************************************************
- * @fn      SendInputSensorControlIndication
- * @brief  This helper function sends input sensor control indication to Sensor Acq task
- * @param  command  - SensorControlCommand_t (currently only SENSOR_CONTROL_SENSOR_ON & SENSOR_CONTROL_SENSOR_OFF supported
- * @param  mask - mask per index to _SensorTable for each sensor to act on
- * @return None
- *
- ***************************************************************************************************/
-void SendInputSensorControlIndication(uint16_t command,	uint16_t mask)        
-{
-    MessageBuffer *pSendMsg = NULLP;
-
-    ASF_assert( ASFCreateMessage( MSG_INPUT_SENSOR_CONTROL_DATA, sizeof(MsgInputSensorsControl), &pSendMsg ) == ASF_OK );
-    pSendMsg->msg.msgInputSensorsControl.Command = command;
-    pSendMsg->msg.msgInputSensorsControl.mask = mask;
-    ASF_assert( ASFSendMessage( SENSOR_ACQ_TASK_ID, pSendMsg ) == ASF_OK );
-}
-
 /****************************************************************************************************
  * @fn	  ActivateResultSensors
  *		  Given a sensor ID, check to be sure that all of the sensors that are needed
